requirements.txt
flask==2.3.2
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0
python-docx==0.8.11
nltk==3.8.1